Created ALPS describing Tent.io hypermedia
With the ALPS(Application-Level Profile Semantics), all the protocols
are consolidated into one protocol so server-to-server communication
and client-to-server communication are more easily implemented.

Using ALPS, the application state is driven by user interactions
using the hypermedia.

This draft is very rough and will still need some work.

A draft also needs to be created for user actions. This will allow
servers to handle notifications similar to Facebook's Open Graph.
Servers can use actions to describe interactions with resources
with the following template:

{user} {action} {target} via {client}

Rendered out the server can notify the user:

titanous starred https://github.com/tent/tent.io via Tent.io for iPhone

The ALPS describing the action needs to instruct servers how to extract
data from the notification and be able to display it to the user.